+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix sentry_logger when SDK is closed from another thread [#1860](https://github.com/getsentry/sentry-ruby/pull/1860)
   - Fixes [#1858](https://github.com/getsentry/sentry-ruby/issues/1858)
 
+- Clear breadcrumbs in `Sentry::TestHelper#teardown_sentry_test` [#1881](https://github.com/getsentry/sentry-ruby/pull/1881)
+
 ## 5.4.1
 
 ### Bug Fixes